Move sff_mgr to a new file
